API overhaul.
1. Change Calc to specify a `map`, run after the fold in `crosstab` and `calc`

Note this allows us to more easily calculate means and similar 2-pass calcs
uniformly for the matrix and the summaries.

2. Encapsulate the rows and cols arrays (levels)

This is because the more common case will be to automatically determine levels
(`crosstab` vs `crosstabWithLevels`). Therefore the call site should not have
to be responsible for these. Also because we want to define a `Table` output
format which needs them.

3. Remove `matrix` from `Value` type. This makes a clearer distinction between
`Crosstab` and `Value` state.

For this reason also, remove `mapCalc`. You never calculate both the matrix
and the summaries.

4. Add `compare`, which is a specialized `map` that takes into account
table, row, col summaries and the previous row and previous column values.