first draft Html view (Crosstab.Html)
  - note: only generates table header, not rows
  - add Window type for dimensions
  - add Display.columnHeaders for generating column header rows
  - add Display.mapValues for concat-mapping LabelledValue to a list of named
    values

Note: the larger question is what "state" does the view module have? Everything
is either derived, or static config. The closest to state that the module could
be responsible for is the Query (controlling the dimensions and their sorting).
But the Crosstab itself should definitely not be put into state here (or
probably anywhere). And it's a bit weird we are editing the query but the
parent module is where the query is used to generate the data to pass back into
the crosstab view.

I see two basic options (to be continued in DIARY.md).