Selectable, Query views: add sorting UI
Note the main thing is that now View.Crosstab.Query has state, which is not (as
you might think) a Query, but parameters to build a Query, which are simplified
for the sorting UI. This means that the Query.update is not needed, and the
parent module (e.g. Selectable) just calls `View.Crosstab.Query.toQuery`
with the value column sorting config and the view query state when it needs
to build the query to run the Crosstab through it.

And connected to this, the value column sorting config is integrated into
the Sortable.Config, specifying the label, render, and way to sort each of
the value columns.

User-specified sorting is not as flexible of course as writing your own
`CompareAxis a` function in code (witness the changes in the Simple example
code), but it's a start that covers the basic cases without overly complicating
the UI.