Selectable: change header rows to render with proper rowspan and colspan
  - move logic to View.Crosstab.Header

Note that this assumes that no funny column sorting has been applied: meaning
that the sorting always compares 'level to level' and thus results in a depth-
first hierarchical overall sort, whatever the sorting within a given dimension.
However, it is technically possible (not in the Sortable UI, but the code
doesn't make it impossible), for a wacky sort to be applied. The rendered
table will not make much sense, and potentially the headers will not match the
values in the table! For this reason, I may put in place something where if it
detects a weird order while pre-computing the row- and colspans, it reverts
to something like the first implementation, with no spanning.

minor: change row headers to th instead of td